supporting API keys
now the Client may be instantiated using a new optional parameter, api_key.
if this parameter is provided, it will always be used in all requests' authorization header instead of performing a POST request to the login endpoint and saving the resulting cookie.

this is as unintrusive as possible in order to maintain backwards compatibility.

requires app-alooma's API key feature in order to work